+++ openreader/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @subsquid/openreader
 
-This log was last generated on Thu, 20 Jan 2022 08:42:53 GMT and should not be manually modified.
+This log was last generated on Tue, 25 Jan 2022 12:44:12 GMT and should not be manually modified.
+
+## 0.4.0
+Tue, 25 Jan 2022 12:44:12 GMT
+
+### Minor changes
+
+- internal: add support for `@index` directives
 
 ## 0.3.3
 Thu, 20 Jan 2022 08:42:53 GMT